Fix sort in GridView; remove redundant code
This commit adds the code for sorting the list of bases to the Grid View instead of it just being in the List View. A redundant sorting option (last modified) is removed because it ran the same code as a different option (last opened). Also, the sorting dropdown was added to each view and workspace instead of being only on the home view. The selected sort option is maintained across different views as they all use the same state (expected behavior).